deal with move and rename of feature.registrater_translation
